ingest/get_eex_data/requirements.txt
boto3
paramiko
python-dateutil